# boardgametracker.client/openapi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "BoardGameTracker API",
    "description": "BoardGameTracker API for managing board game collections and play sessions",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/game/{id}": {
      "delete": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "get": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/gamenight/{id}": {
      "delete": {
        "tags": [
          "GameNight"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/manual/game/{gameId}": {
      "get": {
        "tags": [
          "Manual"
        ],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManualDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManualDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManualDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Manual"
        ],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              },
              "encoding": {
                "Files": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManualDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManualDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManualDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/manual/{id}/reindex": {
      "post": {
        "tags": [
          "Manual"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/manual/{id}": {
      "delete": {
        "tags": [
          "Manual"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/manual/{id}/download": {
      "get": {
        "tags": [
          "Manual"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/manual/{id}/page/{page}/image": {
      "get": {
        "tags": [
          "Manual"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/manual/gamenight/{linkId}": {
      "get": {
        "tags": [
          "Manual"
        ],
        "parameters": [
          {
            "name": "linkId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameNightManualsDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameNightManualsDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameNightManualsDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/manual/gamenight/{linkId}/manual/{manualId}/download": {
      "get": {
        "tags": [
          "Manual"
        ],
        "parameters": [
          {
            "name": "linkId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "manualId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExpansionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "bggId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "gameId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GameDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "yearPublished": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "shopUrl": {
            "type": "string",
            "nullable": true
          },
          "changeDetectionWatchId": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "minPlayers": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxPlayers": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "minPlayTime": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxPlayTime": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "minAge": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isLoaned": {
            "type": "boolean"
          },
          "rating": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "weight": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "bggId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/GameState"
          },
          "hasScoring": {
            "type": "boolean"
          },
          "buyingPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "soldPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "additionDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "expansions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpansionDto"
            },
            "nullable": true
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameLinkDto"
            },
            "nullable": true
          },
          "mechanics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameLinkDto"
            },
            "nullable": true
          },
          "people": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GamePersonDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GameLinkDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "GameNightManualsDto": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "gameTitle": {
            "type": "string"
          },
          "manuals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManualDto"
            }
          }
        },
        "additionalProperties": false
      },
...
      "GamePersonDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/PersonType"
          }
        },
        "additionalProperties": false
      },
...
      "GameState": {
        "enum": [
          "wanted",
          "owned",
          "previouslyOwned",
          "forTrade"
        ],
        "type": "string"
      },
...
      "ManualDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "fileSizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "uploadDate": {
            "type": "string",
            "format": "date-time"
          },
          "contentType": {
            "type": "string"
          },
          "indexStatus": {
            "$ref": "#/components/schemas/ManualIndexStatus"
          },
          "indexedChunkCount": {
            "type": "integer",
            "format": "int32"
          },
          "indexError": {
            "type": "string",
            "nullable": true
          },
          "indexedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ManualIndexStatus": {
        "enum": [
          "pending",
          "indexing",
          "indexed",
          "failed"
        ],
        "type": "string"
      },
...
      "PersonType": {
        "enum": [
          "designer",
          "artist",
          "publisher"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Call POST /api/auth/login and the token is captured automatically, or paste a JWT here.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}